extras: one home for FTS5 query handling (dirge-fmuu)
Three call sites carried independent FTS5 sanitizers with three threat
models: session_search's 6-step Hermes pipeline for user queries,
memory_db's token quoting for model text, and the cross-session
extractor's alphanumeric-only OR-join for seed terms. A syntax hazard
fixed in one stayed open in the others — the bare-apostrophe FTS5
error already bit once (dirge-6js7).

extras::fts now owns all three (sanitize_query / quote_terms /
or_query) with a module doc saying which to use for which input shape.
Pure moves, no behavior changes; the pure-function tests moved along
with two extra edge cases.
