+++ .beads/issues.jsonl
@@ -89,7 +89,7 @@
 {"_type":"issue","id":"dirge-9f1","title":"Chat history ignores 120-col content_width cap","description":"max_line_width and wrap_line use raw content_cols, so on wide terminals scrollback overflows the centered band into divider/panel margin.","status":"closed","priority":1,"issue_type":"bug","assignee":"Yogthos","owner":"yogthos@gmail.com","created_at":"2026-05-21T22:17:33Z","created_by":"Yogthos","updated_at":"2026-05-21T22:26:36Z","started_at":"2026-05-21T22:17:42Z","closed_at":"2026-05-21T22:26:36Z","close_reason":"Closed","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"dirge-woq","title":"R1: fix 3 critical plugin bugs (FFI panic, dialog deadlock, init hang)","description":"From the plugin subsystem audit: (1) wrap JanetCFunctions in catch_unwind so Rust panics don't unwind across the C-FFI boundary into Janet; (2) cancel send_dialog's reply_rx.recv() on worker shutdown so the worker thread doesn't block forever when the UI exits mid-dialog; (3) add timeout to the init handshake so a worker panic before init_tx.send() doesn't hang the main thread. Also: (4) bounds-assert wrap_string's i32 cast for the unlikely \u003e2GB case, (5) make take_string_slot atomic to close the race window, (6) don't eat unrelated user events in the dialog arm.","status":"closed","priority":1,"issue_type":"bug","owner":"yogthos@gmail.com","created_at":"2026-05-20T14:59:57Z","created_by":"Yogthos","updated_at":"2026-05-20T15:30:28Z","started_at":"2026-05-20T15:00:10Z","closed_at":"2026-05-20T15:30:28Z","dependency_count":0,"dependent_count":1,"comment_count":0}
 {"_type":"issue","id":"dirge-fmuu","title":"Single FTS5 query sanitizer","description":"Three independent implementations: session_search sanitize_fts5_query (6-step), memory_db fts_quote (token quoting), cross_session build_fts_or_query (alphanumeric-only OR-join). Security-sensitive triplication. Consolidate on one sanitizer with thin adapters for OR-join semantics.","status":"open","priority":2,"issue_type":"task","owner":"yogthos@gmail.com","created_at":"2026-06-10T03:23:04Z","created_by":"Yogthos","updated_at":"2026-06-10T03:23:04Z","dependency_count":0,"dependent_count":0,"comment_count":0}
-{"_type":"issue","id":"dirge-zact","title":"Consolidate 4x forked-runner drain loops in review.rs","description":"run_background_review / run_curator_review / run_memory_curator_review / run_cross_session_extraction copy-paste the AgentRunner destructure + AbortRunnerOnDrop + event drain with drifted error handling. Extract one drain_forked_runner helper; slot rollback stays explicit at the background-review call site.","status":"open","priority":2,"issue_type":"task","owner":"yogthos@gmail.com","created_at":"2026-06-10T03:23:04Z","created_by":"Yogthos","updated_at":"2026-06-10T03:23:04Z","dependency_count":0,"dependent_count":0,"comment_count":0}
+{"_type":"issue","id":"dirge-zact","title":"Consolidate 4x forked-runner drain loops in review.rs","description":"run_background_review / run_curator_review / run_memory_curator_review / run_cross_session_extraction copy-paste the AgentRunner destructure + AbortRunnerOnDrop + event drain with drifted error handling. Extract one drain_forked_runner helper; slot rollback stays explicit at the background-review call site.","status":"closed","priority":2,"issue_type":"task","assignee":"Yogthos","owner":"yogthos@gmail.com","created_at":"2026-06-10T03:23:04Z","created_by":"Yogthos","updated_at":"2026-06-10T03:45:34Z","started_at":"2026-06-10T03:41:02Z","closed_at":"2026-06-10T03:45:34Z","close_reason":"drain_forked_runner extracted; four call sites are thin wrappers; fake-runner tests cover order, first-error, channel-close","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"dirge-18v2","title":"Headless JSON output claims success on abnormal end","description":"provider/run.rs hardcodes is_error:false + subtype:success. Error events return early (fine), but channel-close-without-Done (runner panic/abort) emits success JSON with partial response, and max-turns truncation emits success with only a stderr notice. Track terminal state and emit is_error/subtype accordingly.","status":"closed","priority":2,"issue_type":"bug","assignee":"Yogthos","owner":"yogthos@gmail.com","created_at":"2026-06-10T03:23:03Z","created_by":"Yogthos","updated_at":"2026-06-10T03:41:01Z","started_at":"2026-06-10T03:36:22Z","closed_at":"2026-06-10T03:41:01Z","close_reason":"RunEnd classification + headless_result_json; incomplete runs exit non-zero","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"dirge-e8sb","title":"Background shell stuck Running when drain task dies","description":"store.finish() only fires from inside the detached drain loop; if that task is aborted or panics, PgKillGuard kills the process but the registry entry stays Running forever — /tasks and bash_output report a dead shell as live. Add a finish-on-drop guard.","status":"closed","priority":2,"issue_type":"bug","assignee":"Yogthos","owner":"yogthos@gmail.com","created_at":"2026-06-10T03:23:02Z","created_by":"Yogthos","updated_at":"2026-06-10T03:36:20Z","started_at":"2026-06-10T03:33:32Z","closed_at":"2026-06-10T03:36:20Z","close_reason":"finish-on-drop backstop in spawn_streaming_shell; abort test covers the path","dependency_count":0,"dependent_count":0,"comment_count":0}
 {"_type":"issue","id":"dirge-slj2","title":"messages_ad FTS delete trigger corrupts index on message delete (latent)","description":"v6 made messages_fts hold a redacted+concatenated projection but the surviving AFTER DELETE trigger issues FTS5 'delete' with raw old.content — external-content FTS5 requires exact indexed values, mismatched deletes corrupt the index. No code deletes messages today (latent). Migration v8: drop messages_ad, document that deletes must clean FTS app-side. Also audit trigram delete path and add warn-log for rows dropped by v6 backfill filter_map.","status":"closed","priority":2,"issue_type":"bug","assignee":"Yogthos","owner":"yogthos@gmail.com","created_at":"2026-06-10T03:23:01Z","created_by":"Yogthos","updated_at":"2026-06-10T03:33:31Z","started_at":"2026-06-10T03:27:19Z","closed_at":"2026-06-10T03:33:31Z","close_reason":"v8 drops messages_ad; delete_session_messages is the safe app-side path; v6 backfill drop-count logged","dependency_count":0,"dependent_count":0,"comment_count":0}
